+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+## 0.1.2
+
+- Added optional logger for ResponseBuilder to enable logging for unexpected errors
+
+## 0.1.1
+
+- Moved HeaderExtension to event_authentication
+
 ## 0.1.0
 
 - Initial version.